sh: use the POSIX extended regex for matching os-release/VERSION_ID!
Do not forget to pass the regular expression used for matching
the value of VERSION_ID read from the /etc/os-release file through
the conversion function that turns it from a Perl-compatible one to
a POSIX extended regular expression. Otherwise the shell implementation
of storpool_variant incorrectly passes '(?:...)' to `grep -E` and fails
to recognize e.g. CentOS 8 (or 7, or 6) via /etc/os-release.

Change-Id: I8d4201b3d62741724d11fb4c59d3f6ed0e4cd184