tox.ini: use sp-python2 for the Python 2.x unit tests
There are some Linux distributions that no longer have a Python 2.x
interpreter package at all, so use the StorPool one instead.
This may mean that the default Tox invocation will fail on hosts that
do not have the StorPool Python 2.x interpreter installed, but this is
acceptable for the moment.

Change-Id: Id76abdd768127ca3f0a87d7210ac4bcba8dea035