Add the supported.repo boolean flag
Add the "supported" structure listing various aspects of the StorPool
environment and infrastructure that may or may not be supported for
the different variants. It currently only contains the "repo" flag that
shows whether StorPool maintains a third-party OS packages repository
containing e.g. the sp-python3 and sp-python3-modules packages.

Declare the repository support for all variants except ALMA9, ROCKY9,
CENTOS6, DEBIAN12, and UBUNTU1604. Strictly speaking, StorPool does
still provide CentOS 6.x and Ubuntu 16.04 repositories, but this may
not be the case for much longer and those variants are officially not
supported any more.

Change-Id: I9b186edeb65e0f47c3e1eabc20c341b6ee1ed6f0