rust: get_all_variants{,_in_order}(), enum-iterator
Implement the get_all_variants() and get_all_variants_in_order()
quite useful functions in the Rust implementation.

Make the VariantKind enum implement (well, derive) the Sequence
trait from the enum-iterator crate so that the values may be
processed in the order they are defined. Note that this is
not the same as the order in which the variants are returned by
the get_all_variants_in_order() function.

Change-Id: I1c590cb826901114fb965820578c4a37d632758b